refactor: move test utils into separate files
