+++ cluster_test.cc
@@ -14,7 +14,7 @@
 
 #include "cluster.h"
 #include "persist/rocksdb/storage.h"
-#include "persist/test_utils.h"
+#include "persist/utils/test_utils.h"
 
 namespace google {
 namespace cloud {
