+++ persist/rocksdb/storage_test.cc
@@ -13,7 +13,7 @@
 // limitations under the License.
 
 #include "persist/rocksdb/storage.h"
-#include "persist/test_utils.h"
+#include "persist/utils/test_utils.h"
 
 namespace google {
 namespace cloud {
